Moves GenCIC
