+++ providers/mockprovider_test.go
@@ -24,7 +24,6 @@ import (
 
 	"github.com/lestrrat-go/jwx/v2/jwa"
 	"github.com/lestrrat-go/jwx/v2/jws"
-	"github.com/openpubkey/openpubkey/providers/mocks"
 	"github.com/openpubkey/openpubkey/util"
 	"github.com/stretchr/testify/require"
 )
@@ -37,7 +36,7 @@ func TestMockProviderTest(t *testing.T) {
 	require.NoError(t, err)
 	idtTemplate.ExtraClaims = map[string]interface{}{"sha": "c7d5b5ff9b2130a53526dcc44a1f69ef0e50d003"}
 
-	cic := mocks.GenCIC(t)
+	cic := GenCIC(t)
 	idToken, err := provider.RequestTokens(context.TODO(), cic)
 	require.NoError(t, err)
 	require.NotNil(t, idToken)
