+++ providers/test.go
@@ -1,4 +1,4 @@
-package mocks
+package providers
 
 import (
 	"testing"
